test: add timely step duration metric test
Add a test to verify that both compute and storage report the
`mz_timely_step_duration_seconds` metric with proper `cluster` and
`worker_id` labels.

Also:
- Move the `Metrics` helper class from `test/cluster/mzcompose.py` to
  `misc/python/materialize/mzcompose/helpers/metrics.py` for reuse
- Use a generic help string for the metric to avoid Prometheus
  registration conflicts between compute and storage
